Conflicts:
	app/controllers/Notes.java
	app/models/Note.java
	app/views/notes/show.scala.html
	app/views/quickCreateForm.scala.html

+++ app/models/Note.java
@@ -1,13 +1,19 @@
 package models;
 
 import java.util.*;
+
+import javax.persistence.*;
+
+import com.avaje.ebean.Expr;
+
 import play.db.ebean.*;
 import play.data.validation.Constraints.*;
 import play.data.format.Formats.*;
 import play.Logger;
 import javax.persistence.*;
-import java.net.MalformedURLException;
 import securesocial.core.java.SecureSocial;
+
+import helpers.UnauthorizedException;
 
 @Entity
 public class Note extends Model {
@@ -24,10 +30,17 @@ public class Note extends Model {
   	@Column(columnDefinition = "TEXT")
 	public String content;
 
-	public String author;
+	@ManyToOne
+  public User author;
 
 	@ManyToMany(cascade = CascadeType.REMOVE)
 	public List<Tag> tags = new ArrayList<Tag>();
+
+  @Column(name = "created_at")
+  public Date createdAt;
+ 
+  @Column(name = "updated_at")
+  public Date updatedAt;
 
 	@OneToMany(mappedBy="note", cascade=CascadeType.ALL)
 	public List<Comment> comments = new ArrayList<Comment>();
@@ -40,30 +53,32 @@ public class Note extends Model {
 	@JoinTable(name="up_votes")
 	public List<User> upVotes = new ArrayList<User>();
 
+  // Needed for ebean finder to be able to order notes by rating. Should not be needed since upVotes and downVotes exist
+  public int rating;
+
 	@ManyToMany(cascade=CascadeType.REMOVE)
 	@JoinTable(name="down_votes")
 	public List<User> downVotes = new ArrayList<User>();
 
 	public Note(String title, User author) {
-		Logger.debug("HEJHEJ " + author);
 		this.title = title;
-		this.author = author.id;
+		this.author = author;
 	}
 
 	public Note(String title, String content, User author) {
-		Logger.debug("HEJHEJ " + author);
 		this.title = title;
 		this.content = content;
-		this.author = author.id;
+		this.author = author;
 	}
 
 	public static Finder<Long, Note> find = new Finder(Long.class, Note.class);
 
-  	public static List<Note> notesBy(User author) {
-      return find.where()
-          .eq("author", author.id)
-          .findList();
-  	}
+
+  public static List<Note> notesBy(User author) {
+    return find.where()
+        .eq("author", author.id)
+        .findList();
+  }
     
 	public static List<Note> all() {
 		return find.all();
@@ -74,9 +89,11 @@ public class Note extends Model {
 		return note;
 	}
 
-	public static Note create(Note note, String tagsList, S3File image) {
+	public static Note create(Note note, String tagsList, User author, S3File image) {
+    note.author = author;
     note.save();
 		if(tagsList != null && !tagsList.equals("") && !tagsList.equals(" ")) {
+      note.tags.clear();
 			note.tags.addAll(Tag.createOrFindAllFromString(tagsList));
 			note.saveManyToManyAssociations("tags");
 		}
@@ -95,9 +112,12 @@ public class Note extends Model {
 
 	public static void delete(Long id) {
 		Note note = find.ref(id);
-		note.delete();
-
-		Tag.clean();
+    if(note.author.equals(User.currentUser())) {
+		  note.delete();
+	   	Tag.clean();
+    } else {
+      throw new UnauthorizedException(User.currentUser(), "delete");
+    }
 	}
 
 	public Note addComment(String content, User author) {
@@ -105,72 +125,84 @@ public class Note extends Model {
     this.comments.add(comment);
     this.save();
     return this;
-    }
+  }
 
-    public void toggleUpVote(User user){
-    	if(user == null){
-    		// Do nothing
-    		Logger.debug("Vote: User is null!");
-    	}
-    	else if(upVotes.contains(user)){
-    		Logger.debug("Vote: removed upvote!");
-    		upVotes.remove(user);
-    		this.saveManyToManyAssociations("upVotes");
-    	} else {
-    		if(downVotes.contains(user)){
-    			Logger.debug("Vote: removed old downvote!");
-    			downVotes.remove(user);
-    			this.saveManyToManyAssociations("downVotes");
-    		}
-    		Logger.debug("Vote: Added upvote!");
-    		upVotes.add(user);
-    		this.saveManyToManyAssociations("upVotes");
-    	}
-    	Logger.debug("Uservotestatus: " + getVoteStatus(user));
-    	this.save();
-    }
+  public static Note update(Note note, String tagsList) {
+    Note existingNote = find.ref(note.id);
+    existingNote.title = note.title;
+    existingNote.content = note.content;
+    existingNote.save();
+    return note;
+  }
 
-    public void toggleDownVote(User user){
-    	if(user == null){
-    		// Do nothing
-    		Logger.debug("Vote: User is null!");
-    	} else if(downVotes.contains(user)){
-    		Logger.debug("Vote: removed downvote!");
-    		downVotes.remove(user);
-    		this.saveManyToManyAssociations("downVotes");
-    	} else {
-    		if(upVotes.contains(user)){
-    			Logger.debug("Vote: removed old upvote!");
-    			upVotes.remove(user);
-    			this.saveManyToManyAssociations("upVotes");
-    		}
-    		Logger.debug("Vote: Added downvote!");
-    		downVotes.add(user);
-    		this.saveManyToManyAssociations("downVotes");
-    	}
-    	Logger.debug("Uservotestatus: " + getVoteStatus(user));
-    	this.save();
-    }
+  public void toggleUpVote(User user){
+  	if(user == null){
+  		// Do nothing
+  		Logger.debug("Vote: User is null!");
+  	}
+  	else if(upVotes.contains(user)){
+  		Logger.debug("Vote: removed upvote!");
+      rating--;
+  		upVotes.remove(user);
+  		this.saveManyToManyAssociations("upVotes");
+  	} else {
+  		if(downVotes.contains(user)){
+  			Logger.debug("Vote: removed old downvote!");
+        rating++;
+  			downVotes.remove(user);
+  			this.saveManyToManyAssociations("downVotes");
+  		}
+  		Logger.debug("Vote: Added upvote!");
+      rating++;
+  		upVotes.add(user);
+  		this.saveManyToManyAssociations("upVotes");
+  	}
+  	Logger.debug("Uservotestatus: " + getVoteStatus(user));
+  	this.save();
+  }
 
-    public int getVoteStatus(User user){
-    	if(upVotes.contains(user)){
-    		return 1;
-    	} else if (downVotes.contains(user)) {
-    		return -1;
-    	} else {
-    		return 0;
-    	}
-    }
+  public void toggleDownVote(User user){
+    rating--;
+  	if(user == null){
+  		// Do nothing
+  		Logger.debug("Vote: User is null!");
+  	} else if(downVotes.contains(user)){
+  		Logger.debug("Vote: removed downvote!");
+      rating++;
+  		downVotes.remove(user);
+  		this.saveManyToManyAssociations("downVotes");
+  	} else {
+  		if(upVotes.contains(user)){
+  			Logger.debug("Vote: removed old upvote!");
+        rating--;
+  			upVotes.remove(user);
+  			this.saveManyToManyAssociations("upVotes");
+  		}
+  		Logger.debug("Vote: Added downvote!");
+      rating--;
+  		downVotes.add(user);
+  		this.saveManyToManyAssociations("downVotes");
+  	}
+  	Logger.debug("Uservotestatus: " + getVoteStatus(user));
+  	this.save();
+  }
 
-    public int getScore(){
-    	Logger.debug("upVotes: " + upVotes.size());
-    	Logger.debug("downVotes: " + downVotes.size());
+  public int getVoteStatus(User user){
+  	if(upVotes.contains(user)){
+  		return 1;
+  	} else if (downVotes.contains(user)) {
+  		return -1;
+  	} else {
+  		return 0;
+  	}
+  }
 
-    	Logger.debug("upvotes: "+ upVotes + "downvotes:" + downVotes);
-    	return (upVotes.size() - downVotes.size());
-    }
-
-
+  public int getScore(){
+  	Logger.debug("upVotes: " + upVotes.size());
+  	Logger.debug("downVotes: " + downVotes.size());
+  	Logger.debug("upvotes: "+ upVotes + "downvotes:" + downVotes);
+  	return (upVotes.size() - downVotes.size());
+  }
 
 	public String extract(int length) {
 		if(content != null) {
@@ -187,6 +219,47 @@ public class Note extends Model {
     Collections.sort(allNotes, new NoteComparator(note));
     Collections.reverse(allNotes);
     return allNotes;
+  }
+
+  /**
+   * Returns a list of notes related to the search query. 
+   * Will look at both content and titles and order them by rating. 
+   */ 
+  public static List<Note> searchNotes(String query) {
+    List<Note> result = new ArrayList<Note>();
+    for (String word : query.split("\\s")) {
+      List<Note> wordResult = find.where()
+        .or(Expr.ilike("title", "%"+word+"%"), Expr.ilike("content", "%"+word+"%")).orderBy("rating")
+        .findList();
+        for (Note note : wordResult) {
+          if (!result.contains(note))
+            result.add(note);
+        }
+    }
+    Collections.reverse(result);
+    return result;
+  }
+
+  @Override
+  public void save() {
+    createdAt();
+    super.save();
+  }
+ 
+  @Override
+  public void update() {
+    updatedAt();
+    super.update();
+  }
+ 
+  @PrePersist
+  void createdAt() {
+    this.createdAt = this.updatedAt = new Date();
+  }
+ 
+  @PreUpdate
+  void updatedAt() {
+    this.updatedAt = new Date();
   }
 
   public static class NoteComparator implements Comparator<Note> {
